fully into the redesign
